[#71]뷰이름 정리 먼저
