+++ Fleming_App/Fleming_App/Model_Test/HandPoseClassifier_ver4_AR/CameraTest2.swift
@@ -4,15 +4,15 @@
 //
 //  Created by Leo Yoon on 10/28/24.
 //
-import SwiftUI
-
-@main
-struct CameraTest2: App {
-    var body : some Scene {
-        WindowGroup {
-            ContentView2()
-//                .frame(width: UIScreen.main.bounds.width / 2, height: UIScreen.main.bounds.height / 2)
-//                .cornerRadius(10)
-        }
-    }
-}
+//import SwiftUI
+//
+//@main
+//struct CameraTest2: App {
+//    var body : some Scene {
+//        WindowGroup {
+//            ContentView2()
+////                .frame(width: UIScreen.main.bounds.width / 2, height: UIScreen.main.bounds.height / 2)
+////                .cornerRadius(10)
+//        }
+//    }
+//}
